requirements.txt
# Core dependencies
python-telegram-bot==21.0.1
python-dotenv==1.0.1

# Notion integration
notion-client==2.2.1

# Date and time handling
python-dateutil==2.8.2
pytz==2024.1

# Data validation
pydantic==2.5.3

# AI Integration (für intelligente Terminextraktion)
openai==1.97.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov==4.1.0
factory-boy==3.3.0
pytest-benchmark==4.0.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
responses==0.24.1
freezegun==1.4.0
hypothesis==6.97.0
locust==2.23.1

# Development dependencies
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
aiohttp==3.9.3

# Email processing
email-validator==2.1.1